feat(orders): :sparkles: remove product from order
These classes handle the removal of a product from an order for the e-commerce app. The `RemoveProductFromOrderCommand` accepts an order ID and a product ID, and this command is processed by the `RemoveProductFromOrder` handler class. This handler interacts with the `OrderRepository` and `ProductRepository` to remove the product from the order's items. The `Order` domain model's removeProductItem method now returns a `DomainError` in case the requested product is not found in the order.